Move test data parameters to separate file
